+++ CHANGELOG.md
@@ -5,7 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-*N/A*
+### Changed
+- Give `<Button>` same size/padding/border/margin as inputs so they line up nicely side by side
 
 ## [0.17.1]
 ### Changed